much better python interface; works OK; also in Jupyter notebook with progress bar
