+++ common.h
@@ -6,10 +6,9 @@
 
 #include <boost/python.hpp>
 
-#include "pylogger.h"
+#define TOMOGRAPHER_EIGEN_ASSERT_EXCEPTION
+#include <tomographer2/tools/eigen_assert_exception.h>
 
-// must be included at top, before any other Eigen header (sets eigen_assert() macro)
-// but it uses python logging, so after "pylogger.h"
 #include "eigpyconv.h"
 
 #include <Eigen/Eigen>
@@ -17,10 +16,27 @@
 #include <tomographer2/tools/cxxutil.h>
 #include <tomographer2/tools/loggers.h>
 
+#include "pylogger.h"
+
+
 
 typedef double RealType;
 typedef int CountIntType;
 
+
+
+// Python/C++ logger:
+
+// to allow for debugging
+//extern Tomographer::Logger::FileLogger tpy_logger(std::stderr, Tomographer::Logger::INFO, true /* display_origin */);
+// to be fast
+//extern Tomographer::Logger::VacuumLogger tpy_logger;
+// by default: slow, but really good integration to Python logging
+extern PyLogger tpy_logger;
+
+
+// the type of the C++ logger we are using -- for any of the above
+typedef decltype(tpy_logger) TPyLoggerType;
